# Added Min and Max loops