--- a/src/ast.rs
+++ b/src/ast.rs
@@ -208,6 +208,8 @@ pub enum Expression {
     For(Box<For>),
     ForN(Box<ForN>),
     Sum(Box<ForN>),
+    Min(Box<ForN>),
+    Max(Box<ForN>),
     If(Box<If>),
     Compare(Box<Compare>),
     UnOp(Box<UnOpExpression>),
@@ -315,6 +317,14 @@ impl Expression {
                     "sum", convert, ignored) {
                 convert.update(range);
                 result = Some(Expression::Sum(Box::new(val)));
+            } else if let Ok((range, val)) = ForN::from_meta_data(
+                    "min", convert, ignored) {
+                convert.update(range);
+                result = Some(Expression::Min(Box::new(val)));
+            } else if let Ok((range, val)) = ForN::from_meta_data(
+                    "max", convert, ignored) {
+                convert.update(range);
+                result = Some(Expression::Max(Box::new(val)));
             } else if let Ok((range, val)) = Loop::from_meta_data(
                     convert, ignored) {
                 convert.update(range);
@@ -367,6 +377,8 @@ impl Expression {
             For(ref for_expr) => for_expr.source_range,
             ForN(ref for_n_expr) => for_n_expr.source_range,
             Sum(ref for_n_expr) => for_n_expr.source_range,
+            Min(ref for_n_expr) => for_n_expr.source_range,
+            Max(ref for_n_expr) => for_n_expr.source_range,
             If(ref if_expr) => if_expr.source_range,
             Compare(ref comp) => comp.source_range,
             UnOp(ref unop) => unop.source_range,