See https://github.com/PistonDevelopers/dyon/issues/119

This PR adds `min` and `max` loops, similar to the ∑ loop.

Example:

```rust
fn main() {
    list := [81, 402, 5, 12, 42]
    min := min i len(list) { list[i] } // prints `min: some([2, 5])`
    max := max i len(list) { list[i] } // prints `max: some([1, 402])`
    println("min: " + to_string(min))
    println("max: " + to_string(max))
}
```

Both loops return `some([arg, val])` where `arg` is the counter argument at the extremal value (arg min/max) and `val` is the value. Returns `none()` if the body did not return any value.

The `min` loop is logically equivalent to:

```rust
fn min(list) -> {
    min_arg := none()
    min_val := none()
    for i len(list) {
        less := if min_arg == none() { true }
                else if list[i] < unwrap(min_val) { true }
                else { false }
        if less {
            min_arg = some(i)
            min_val = some(list[i])
        }
    }
    return if min_arg == none() { none() }
           else { some([unwrap(min_arg), unwrap(min_val)]) }
}
```

The `max` loop is logically equivalent to:

```rust
fn max(list) -> {
    max_arg := none()
    max_val := none()
    for i len(list) {
        more := if max_arg == none() { true }
                else if list[i] > unwrap(max_val) { true }
                else { false }
        if more {
            max_arg = some(i)
            max_val = some(list[i])
        }
    }
    return if max_arg == none() { none() }
           else { some([unwrap(max_arg), unwrap(max_val)]) }
}
```

### Motivation

Min and Max are commonly used forms of loops, but can be tricky to get right on the first try. These new loops are:

- Easy to write and read
- Can be used part of an expression
- Contains the information that is commonly needed (arg min/max)

### Performance

The Min and Max loops are 5.5x faster than the equivalent function.

```
Before:

test tests::bench_add    ... bench:  43,060,552 ns/iter (+/- 6,013,903)
test tests::bench_add_n  ... bench:  15,853,633 ns/iter (+/- 2,861,556)
test tests::bench_array  ... bench:  17,817,021 ns/iter (+/- 2,701,964)
test tests::bench_call   ... bench:  16,915,996 ns/iter (+/- 3,649,586)
test tests::bench_len    ... bench:   3,185,801 ns/iter (+/- 107,785)
test tests::bench_main   ... bench:   1,949,300 ns/iter (+/- 80,327)
test tests::bench_n_body ... bench: 210,557,118 ns/iter (+/- 12,233,201)
test tests::bench_object ... bench:  27,041,261 ns/iter (+/- 4,641,913)
test tests::bench_sum    ... bench:   6,703,750 ns/iter (+/- 647,081)

After:

test tests::bench_add    ... bench:  42,196,013 ns/iter (+/- 6,000,027)
test tests::bench_add_n  ... bench:  15,198,933 ns/iter (+/- 2,980,709)
test tests::bench_array  ... bench:  18,772,144 ns/iter (+/- 2,995,647)
test tests::bench_call   ... bench:  17,803,821 ns/iter (+/- 2,883,615)
test tests::bench_len    ... bench:   3,350,437 ns/iter (+/- 140,630)
test tests::bench_main   ... bench:   2,063,016 ns/iter (+/- 96,271)
test tests::bench_min    ... bench:  15,943,394 ns/iter (+/- 2,471,873) <--- Min loop
test tests::bench_min_fn ... bench:  78,283,883 ns/iter (+/- 6,072,677) <--- Min function
test tests::bench_n_body ... bench: 212,756,276 ns/iter (+/- 8,928,322)
test tests::bench_object ... bench:  26,145,193 ns/iter (+/- 4,503,270)
test tests::bench_sum    ... bench:   6,924,815 ns/iter (+/- 1,116,980)

(bench_min_fn - bench_main) / (bench_min - bench_main)
= (78283883 - 2063016) / (15943394 - 2063016)
= 5.491267384793123
```